pyproject.toml
[project]
name = "emr-icu-handover"
version = "0.1.0"
description = "Multi-agent clinical handoff summarization from heterogeneous EMR Excel inputs"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.43.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.10.0",
    "jupyter>=1.1.1",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.0",
]
